requirements.txt
altgraph==0.17.4
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
packaging==25.0
pefile==2023.2.7
pyaes==1.6.1
pyasn1==0.6.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
python-dotenv==1.1.0
pywin32-ctypes==0.2.3
requests==2.32.3
rsa==4.9.1
setuptools==80.9.0
Telethon==1.40.0
urllib3==2.4.0
fastapi==0.100.0
uvicorn==0.23.1